CMake and bug fix
