+++ headers/WindowManager.hpp
@@ -3,6 +3,7 @@
 #include <vector>
 #include <memory>
 #include <map>
+#include <cstring>
 #include "Window.hpp"
 #include "Utils.h"
 #include "WindowPtr.hpp"
@@ -71,7 +72,11 @@ namespace tmk
         {
             this->render_buffer(id);
 
-            std::wcout << "\x1b[1J\x1b[0;0H" << buffer_;
+            // Need to do this because std::wcout << buffer_ generates weird chars at the end
+            for(size_t i = 0; i < width_ * height_; ++i)
+            {
+                std::wcout << buffer_[i];
+            }
         }
 
         void set_visible(std::shared_ptr<Window> window)
